requirements.txt
Pillow==6.2.0
urllib3==1.25.6
python-dotenv==0.10.3
instabot==0.4.5
requests==2.22.0
